+++ solver_rust_conic/totsu/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Re-export `totsu_core` module.
 ### Changed
 - `Debug` of `MatBuild` prints like `Display`.
 ### Deprecated